--- packages/lix-sdk/src/change-queue/index.ts
@@ -1,1 +1,0 @@
-export { changeQueueSettled } from "./change-queue-settled.js";
