+++ packages/lix-sdk/src/file-queue/index.ts
@@ -1,0 +1,1 @@
+export { fileQueueSettled as changeQueueSettled } from "./file-queue-settled.js";
