+++ packages/lix-sdk/src/index.ts
@@ -2,7 +2,7 @@ export * from "./account/index.js";
 export * from "./change/index.js";
 export * from "./change-set/index.js";
 export * from "./change-conflict/index.js";
-export * from "./change-queue/index.js";
+export * from "./file-queue/index.js";
 export * from "./change-schema/index.js";
 export * from "./database/index.js";
 export * from "./discussion/index.js";
