+++ packages/lix-sdk/src/lix/open-lix.ts
@@ -2,8 +2,8 @@ import type { LixPlugin } from "../plugin/lix-plugin.js";
 import { loadPlugins } from "../plugin/load-plugin.js";
 import { contentFromDatabase, type SqliteDatabase } from "sqlite-wasm-kysely";
 import { initDb } from "../database/init-db.js";
-import { initChangeQueue } from "../change-queue/init-change-queue.js";
-import { changeQueueSettled } from "../change-queue/change-queue-settled.js";
+import { initChangeQueue } from "../file-queue/file-queue-process.js";
+import { fileQueueSettled } from "../file-queue/file-queue-settled.js";
 import type { Kysely } from "kysely";
 import type { LixDatabaseSchema } from "../database/schema.js";
 import { initSyncProcess } from "../sync/sync-process.js";
@@ -78,7 +78,7 @@ export async function openLix(args: {
 	};
 
 	const toBlob = async () => {
-		await changeQueueSettled({ lix: { db } });
+		await fileQueueSettled({ lix: { db } });
 		return new Blob([contentFromDatabase(args.database)]);
 	};
 
@@ -95,7 +95,7 @@ export async function openLix(args: {
 		toBlob,
 		plugin,
 		close: async () => {
-			await changeQueueSettled({ lix: { db } });
+			await fileQueueSettled({ lix: { db } });
 			// args.database.close();
 			// await db.destroy();
 		},
